Fixed location parser * Fixed location parsing double as integer.

+++ src/main/java/xyz/theprogramsrc/lastlocation/hooks/LastLocationAction.java
@@ -72,7 +72,7 @@ public class LastLocationAction implements SuperAuthAction, Listener {
         if(s.length < 6){
             return null;
         }else{
-            return new Location(Bukkit.getWorld(s[0]), Integer.parseInt(s[1]), Integer.parseInt(s[2]), Integer.parseInt(s[3]), Float.parseFloat(s[4]), Float.parseFloat(s[5]));
+            return new Location(Bukkit.getWorld(s[0]), Double.parseDouble(s[1]), Double.parseDouble(s[2]), Double.parseDouble(s[3]), Float.parseFloat(s[4]), Float.parseFloat(s[5]));
         }
     }
 }
